requirements.txt
# Web アプリケーションフレームワーク
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.29.0

# データベース関連
sqlalchemy>=2.0.0

# データモデリング
pydantic>=2.0.0

# スクレイピング関連
requests>=2.31.0
beautifulsoup4>=4.12.0

# 日付処理
python-dateutil>=2.8.2

# ロギング（標準ライブラリ）
logging  # Pythonに標準搭載

# 開発ツール（オプション）
black>=23.0.0  # コードフォーマッター
flake8>=6.1.0  # リンター
pytest>=7.4.0  # テストフレームワーク

# 既存の依存関係に追加
psycopg2-binary>=2.9.9  # PostgreSQLアダプター
python-dotenv>=1.0.0    # 環境変数管理
